Return a 401 when attempting to use tokenized login with a foreign org.
The authentication middleware already checks organization membership so
this wasn’t a security hole, you’d be thrown out to the login page by
the redirect request. Still, for the sake of correctness the tokenized
login request should itself do a 401 in these circumstances.

+++ src/php/FrontEnd/App/Controller/Organization/TokenizedLoginController.php
@@ -28,6 +28,7 @@ class TokenizedLoginController
 
     public function getAction(Request $request): Response
     {
+        $organization = $request->attributes->get('organization');
         $token = $request->query->get('t');
 
         if (null == $token) {
@@ -40,6 +41,10 @@ class TokenizedLoginController
         }
 
         $person = $this->personRepository->findById($tokenizedLogin['person_id']);
+        if ($organization->getId() !== $person['organization_id']) {
+            return new Response(null, 401);
+        }
+
         $this->tokenizedLoginRepository->deleteAllForPerson($person['id']);
 
         $session = $request->getSession();
